refactor: 오타 수정 및 리라우트 위치 변경
