+++ apps/admin/pages/seller/account.tsx
@@ -16,17 +16,20 @@ export function SellerAccountList(): JSX.Element {
   const router = useRouter();
   const toast = useToast();
 
+  if (!isLoading && !['super', 'full'].includes(profile.adminClass)) {
+    toast({
+      title: '권한없는 계정',
+      status: 'error',
+    });
+    router.push('/admin');
+  }
+
   useEffect(() => {
-    if (!isLoading && !['super', 'full'].includes(profile.adminClass)) {
-      toast({
-        title: '권한없는 계정',
-        status: 'error',
-      });
-      router.push('/admin');
-    } else if (!isLoading && ['super', 'full'].includes(profile.adminClass)) {
-      mutateAsync({ infoType: 'sellerBusinessRegistration', actionType: 'view' });
+    if (!isLoading && ['super', 'full'].includes(profile.adminClass)) {
+      mutateAsync({ infoType: 'sellerSettlementAccount', actionType: 'view' });
     }
   }, []);
+
   return (
     <AdminPageLayout>
       <Box position="relative">
